+++ test/client.test.ts
@@ -1,4 +1,5 @@
 import { createServer, AddressInfo, Socket } from 'net';
+import { describe, expect, jest, test } from '@jest/globals';
 import { testWithClient } from './helper';
 import {
     Client,